refactor: move to common
