+++ packages/cli/package.json
@@ -30,8 +30,8 @@
     "test:ci": "pnpm run test"
   },
   "dependencies": {
-    "@aws-sdk/client-kms": "*",
-    "@ethersproject/bytes": "^5.5.4",
+    "@ethersproject/abi": "^5.7.0",
+    "@ethersproject/providers": "^5.7.2",
     "@improbable-eng/grpc-web": "^0.15.0",
     "@improbable-eng/grpc-web-node-http-transport": "^0.15.0",
     "@latticexyz/abi-ts": "workspace:*",
@@ -45,13 +45,11 @@
     "@latticexyz/utils": "workspace:*",
     "@latticexyz/world": "workspace:*",
     "@latticexyz/world-modules": "workspace:*",
-    "asn1.js": "^5.4.1",
     "chalk": "^5.0.1",
     "chokidar": "^3.5.3",
     "debug": "^4.3.4",
     "dotenv": "^16.0.3",
     "ejs": "^3.1.8",
-    "ethereumjs-util": "^7.1.5",
     "ethers": "^5.7.2",
     "execa": "^7.0.0",
     "glob": "^8.0.3",
