+++ packages/common/src/account/kms.ts
@@ -1,7 +1,7 @@
+import { GetPublicKeyCommand, GetPublicKeyCommandOutput, SignCommand, SignCommandOutput } from "@aws-sdk/client-kms";
 import { SignParams, GetEthAddressFromKMSparams, GetPublicKeyParams } from "./types";
 import { getEthAddressFromPublicKey } from "./eth";
 import { utils } from "ethers";
-import { GetPublicKeyCommand, GetPublicKeyCommandOutput, SignCommand, SignCommandOutput } from "@aws-sdk/client-kms";
 import { Address } from "viem";
 
 export const getPublicKey = (getPublicKeyParams: GetPublicKeyParams): Promise<GetPublicKeyCommandOutput> => {
@@ -11,11 +11,13 @@ export const getPublicKey = (getPublicKeyParams: GetPublicKeyParams): Promise<Ge
   return kmsInstance.send(command);
 };
 
-export const getEthAddressFromKMS = async (getEthAddressFromKMSparams: GetEthAddressFromKMSparams) => {
+export const getEthAddressFromKMS = async (
+  getEthAddressFromKMSparams: GetEthAddressFromKMSparams,
+): Promise<Address> => {
   const { keyId, kmsInstance } = getEthAddressFromKMSparams;
   const KMSKey = await getPublicKey({ keyId, kmsInstance });
 
-  return getEthAddressFromPublicKey(KMSKey.PublicKey) as Address;
+  return getEthAddressFromPublicKey(KMSKey.PublicKey as Uint8Array) as Address;
 };
 
 export const sign = async (signParams: SignParams): Promise<SignCommandOutput> => {
