+++ packages/common/src/account/kmsAccount.ts
@@ -1,10 +1,10 @@
-import { hashMessage, hashTypedData, signatureToHex } from "viem";
-import { toAccount } from "viem/accounts";
 import { KMSClient } from "@aws-sdk/client-kms";
+import { LocalAccount, hashMessage, hashTypedData, signatureToHex } from "viem";
+import { toAccount } from "viem/accounts";
 import { getEthAddressFromKMS } from "./kms";
 import { createSignature } from "./eth";
 
-export async function createKmsAccount(keyId: string, kmsInstance: KMSClient) {
+export async function createKmsAccount(keyId: string, kmsInstance: KMSClient): Promise<LocalAccount> {
   const address = await getEthAddressFromKMS({ kmsInstance, keyId });
 
   const account = toAccount({
